+++ restygwt/src/main/java/org/fusesource/restygwt/client/cache/QueuableRuntimeCacheStorage.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2009-2010 the original author or authors. See the notice.md file distributed with
  * this work for additional information regarding copyright ownership.
- *
+ * 
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
- *
+ * 
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software distributed under the License
  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
  * or implied. See the License for the specific language governing permissions and limitations under
@@ -23,6 +23,7 @@ import java.util.logging.Logger;
 
 import org.fusesource.restygwt.client.Defaults;
 
+import com.google.gwt.core.client.GWT;
 import com.google.gwt.http.client.RequestCallback;
 import com.google.gwt.http.client.Response;
 import com.google.gwt.user.client.Timer;
@@ -32,15 +33,14 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
     /**
      * how long will a cachekey be allowed to exist
      */
-    public long DEFAULT_LIFETIME_MS = 90 * 1000;
+    public int DEFAULT_LIFETIME_MS = 90 * 1000;
 
     /**
      * key <> value hashmap for holding cache values. nothing special here.
-     *
+     * 
      * invalidated values will be dropped by timer
      */
-    protected final Map<CacheKey, Response> cache =
-            new HashMap<CacheKey, Response>();
+    protected final Map<CacheKey, Response> cache = new HashMap<CacheKey, Response>();
 
     /**
      * keep all callbacks to a cachekey that need to be performed when a response comes from the
@@ -58,8 +58,7 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
     /**
      * reference from scope to cachekey
      */
-    protected final Map<String, List<CacheKey>> scopeRef =
-        new HashMap<String, List<CacheKey>>();
+    protected final Map<String, List<CacheKey>> scopeRef = new HashMap<String, List<CacheKey>>();
 
     public Response getResultOrReturnNull(CacheKey key) {
         return cache.get(key);
@@ -70,9 +69,21 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
         putResult(key, response, "");
     }
 
+    @Override
+    public void putResult(final CacheKey key, final Response response, final int lifetime) {
+        putResult(key, response, lifetime, "");
+    }
+
     public void putResult(final CacheKey key, final Response response, final String scope) {
+        putResult(key, response, DEFAULT_LIFETIME_MS, scope);
+    }
+
+    public void putResult(final CacheKey key, final Response response, final int lifetime,
+            final String scope) {
+        GWT.log("put result: " + key + " with lifetime: " + lifetime);
         final Timer t = new Timer() {
             public void run() {
+                GWT.log("remove result: " + key);
                 removeResult(key, scope);
             }
         };
@@ -81,29 +92,34 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
 
         getScope(scope).add(key);
 
-        t.schedule((int) DEFAULT_LIFETIME_MS);
+        t.schedule(lifetime);
         timers.add(t);
     }
 
     /**
      * put a result to one or many scopes
-     *
+     * 
      * e.g. <code>
      *      putResult("foo_key", "bar_result", {"UserDto", "ProfileDto"}) {
      * </code>
      */
     @Override
     public void putResult(final CacheKey key, final Response response, final String[] scopes) {
+        putResult(key, response, DEFAULT_LIFETIME_MS, scopes);
+    }
+
+    @Override
+    public void putResult(CacheKey key, Response response, int lifetime, String[] scopes) {
         if (null == scopes) {
-            putResult(key, response);
+            putResult(key, response, lifetime);
             return;
         }
 
         int count = 0;
 
-        for (String s: scopes) {
+        for (String s : scopes) {
             if (0 == count) {
-                putResult(key, response, s);
+                putResult(key, response, lifetime, s);
             } else {
                 getScope(s).add(key);
             }
@@ -125,9 +141,8 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
                 }
 
                 if (hasCallback(key)) {
-                    Logger.getLogger(QueuableRuntimeCacheStorage.class.getName())
-                            .warning("cache-key " + key + " still has "
-                                    + pendingCallbacks.get(key).size()
+                    Logger.getLogger(QueuableRuntimeCacheStorage.class.getName()).warning(
+                            "cache-key " + key + " still has " + pendingCallbacks.get(key).size()
                                     + " callbacks, but will be removed now.");
                 }
                 Logger.getLogger(QueuableRuntimeCacheStorage.class.getName()).finer(
@@ -207,7 +222,7 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
             if (Defaults.canLog()) {
                 Logger.getLogger(QueuableRuntimeCacheStorage.class.getName()).fine(
                         "will remove " + currentScope.size() + " elements from cache, scope: "
-                        + scope);
+                                + scope);
             }
 
             for (CacheKey k : currentScope) {
@@ -231,7 +246,7 @@ public class QueuableRuntimeCacheStorage implements ScopableQueueableCacheStorag
 
     /**
      * helper method to access a scope
-     *
+     * 
      * @param scope
      * @return
      */