+++ restygwt/src/main/java/org/fusesource/restygwt/client/cache/ScopableCacheStorage.java
@@ -2,13 +2,13 @@
  * Copyright (C) 2009-2010 the original author or authors.
  * See the notice.md file distributed with this work for additional
  * information regarding copyright ownership.
- *
+ * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,12 +17,13 @@
  */
 package org.fusesource.restygwt.client.cache;
 
+import com.google.gwt.http.client.Response;
 
 public interface ScopableCacheStorage<T> extends CacheStorage<T> {
 
     /**
      * put by ident/scope. e.g. to invalidate later on by domain class
-     *
+     * 
      * @param key
      * @param scope
      * @param response
@@ -31,7 +32,7 @@ public interface ScopableCacheStorage<T> extends CacheStorage<T> {
 
     /**
      * put by ident/scope. e.g. to invalidate later on by domain class
-     *
+     * 
      * @param key
      * @param scope
      * @param response
@@ -39,7 +40,19 @@ public interface ScopableCacheStorage<T> extends CacheStorage<T> {
     public void putResult(final CacheKey key, final T response, final String[] scopes);
 
     /**
+     * put by ident/scope. e.g. to invalidate later on by domain class
+     * 
+     * @param key
+     * @param scope
+     * @param response
+     */
+    public void putResult(final CacheKey key, final T response, final int lifetime,
+            final String[] scopes);
+
+    /**
      * purge a particular ident, e.g. domain scope
      */
     public void purge(final String scope);
+
+    void putResult(CacheKey key, Response response, int lifetime);
 }